update codes, fix bug
