+++ example/Blinker_AHRS/AHRS_WiFi/AHRS_WiFi.py
@@ -6,6 +6,8 @@ Blinker.attachAhrs()
 
 if __name__ == '__main__':
     while True:
+        Blinker.run()
+        
         if Blinker.available() == True:
             BLINKER_LOG('Blinker.readString(): ', Blinker.readString())
             Blinker.print(Blinker.times())
