+++ example/Blinker_Joystick/Joystick_WiFi/Joystick_WiFi.py
@@ -5,6 +5,8 @@ Blinker.begin()
 
 if __name__ == '__main__':
     while True:
+        Blinker.run()
+        
         if Blinker.available() == True:
             BLINKER_LOG('Blinker.readString(): ', Blinker.readString())
             Blinker.print(Blinker.times())
