+++ example/Blinker_Notify/Notify_WiFi/Notify_WiFi.py
@@ -8,6 +8,8 @@ Blinker.wInit(BUTTON_1, W_BUTTON)
 
 if __name__ == '__main__':
     while True:
+        Blinker.run()
+        
         if Blinker.available() == True:
             BLINKER_LOG('Blinker.readString(): ', Blinker.readString())
             Blinker.print(Blinker.times())
