+++ example/Blinker_RGB/RGB_WiFi/RGB_WiFi.py
@@ -8,6 +8,8 @@ Blinker.wInit(RGB_1, W_RGB)
 
 if __name__ == '__main__':
     while True:
+        Blinker.run()
+        
         if Blinker.available() == True:
             BLINKER_LOG('Blinker.readString(): ', Blinker.readString())
             Blinker.print(Blinker.times())
