+++ example/Blinker_Slider/Slider_WiFi/Slider_WiFi.py
@@ -8,6 +8,8 @@ Blinker.wInit(SLIDER_1, W_SLIDER)
 
 if __name__ == '__main__':
     while True:
+        Blinker.run()
+        
         if Blinker.available() == True:
             BLINKER_LOG('Blinker.readString(): ', Blinker.readString())
             Blinker.print(Blinker.times())
