+++ example/Blinker_Toggle/Toggle_WiFi/Toggle_WiFi.py
@@ -6,6 +6,8 @@ Blinker.wInit(TOGGLE_1, W_TOGGLE)
 
 if __name__ == '__main__':
     while True:
+        Blinker.run()
+        
         if Blinker.available() == True:
             BLINKER_LOG('Blinker.readString(): ', Blinker.readString())
             Blinker.print(Blinker.times())
